+++ apps/backend/src/referentiels/switch-to-te/merge-commentaires/merge-commentaires.rules.ts
@@ -1,4 +1,4 @@
-import { sortByReferentielOrder } from '../shared/origine-resolution';
+import { sortByReferentielOrder } from '../shared/origine.rules';
 import {
   getScoreRatios,
   getStatutAvancement,
